requirements.txt

llama-index 
pypdf 
python-dotenv 
pandas 

llama-index-llms-ollama
